backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.4
diskcache==5.6.1
fastapi==0.100.0
h11==0.14.0
idna==3.4
llama-cpp-python==0.1.69
numpy==1.25.1
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
python-dotenv==1.0.0
sniffio==1.3.0
sse-starlette==1.6.1
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
